refactor(vortex-velox): Separate visitor implementation
